Rearrange files
Instead of having all routing handlers inside index.js, I
split them up into their own directories and require them
instead now. This rearrangement required I import a few functions
I needed in both places, so I created functions.js to export them.

- Moved all routes from index.js into routes/
- Added middleware dir
- Added functions.js
